TI:armv7: Re-order slightly the generic CONFIG options, expand related comments
Re-group the general options to note things that can be removed safely
to reduce binary size when not required, and expand the comment about
what commands we do include to note it could be replaced with a specific
list of required one.  While at it, move the CMD parts of various
blocks IP to the end of the list for consistency.

Signed-off-by: Tom Rini <trini@ti.com>