Makefile: Do not show make debug messages
In commit f85dc95, the top Makefile was adjusted to the new
boards.cfg format.

But at the same time, -d option was added.

If you configure and make separately, for example
like follows:

    make omap4_panda_config
    make CROSS_COMPILE=<your_compiler_prefix>

it works fine as it did before.

But if you do them in one time like follows:

    make omap4_panda CROSS_COMPILE=<your_compiler_prefix>

The log is sprinkled with annoying make debug messages.

This commit deletes -d option again.

Signed-off-by: Masahiro Yamada <yamada.m@jp.panasonic.com>
Tested-by: Nishanth Menon  <nm@ti.com>